+++ stagemonitor-logging/src/main/java/org/stagemonitor/logging/LoggingPlugin.java
@@ -1,0 +1,14 @@
+package org.stagemonitor.logging;
+
+import com.codahale.metrics.MetricRegistry;
+import org.stagemonitor.core.Configuration;
+import org.stagemonitor.core.StageMonitorPlugin;
+import org.stagemonitor.core.rest.RestClient;
+
+public class LoggingPlugin implements StageMonitorPlugin {
+
+	@Override
+	public void initializePlugin(MetricRegistry metricRegistry, Configuration configuration) {
+		RestClient.sendGrafanaDashboardAsync(configuration.getElasticsearchUrl(), "Logging.json");
+	}
+}