+++ CHANGELOG.md
@@ -21,3 +21,7 @@
  * annotate methods with @MonitorRequests or make RequestMonitorAspect concrete in aop.xml to monitor method executions like RMI calls
  * support for metrics-annotations: @Timed, @Metered, @ExceptionMetered, @Gauge (enabled by @MonitorGauges on type)
  * added copyToLib gradle task. to assemble all libs necessary e.g. to run staemonitor-spring-mvc, execute `./gradlew copyToJar`. Then go to stagemonitor-spring-mvc/build/libs/lib
+ * Logging Plugin
+   * MeterLoggingAspect: Creates meters for trace, debug, info, warn, error and fatal and updates them on every log statement
+   * TimeLoggingAspect: Creates timers for trace, debug, info, warn, error and fatal and updates them on every log statement
+   * InformationAppendingLoggingAspect: Appends the requestId, application name, instance name and host name to line to each logging statement